LayoutTests/imported/w3c/web-platform-tests/html/user-activation/activation-trigger-mouse-left.html

<!DOCTYPE html>
<html>
  <head>
    <title>Tests user activation from a mouse click</title>
    <link
      rel="help"
      href="https://html.spec.whatwg.org/multipage/interaction.html#activation-triggering-input-event"
    />
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <script src="/resources/testdriver.js"></script>
    <script src="/resources/testdriver-vendor.js"></script>
    <script src="resources/utils.js"></script>
  </head>
  <body>
    <h1>Just a h1</h1>
  </body>
  <script>
    promise_test(async () => {
      const mousedown_event = getEvent("mousedown");
      const mouseup_event = getEvent("mouseup");
      const click_event = getEvent("click");

      await test_driver.click(document.querySelector("h1"));

      await mousedown_event;
      assert_true(
        navigator.userActivation.isActive,
        "mousedown event should result in activation"
      );
      await consumeTransientActivation();

      await mouseup_event;
      assert_false(
        navigator.userActivation.isActive,
        "mouseup should have no activation after mousedown consumption"
      );

      await click_event;
      assert_false(
        navigator.userActivation.isActive,
        "click should have no activation after mousedown consumption"
      );
    }, "Activation through left-click mouse event");
  </script>
</html>
